fix: refactor based on comments
