+++ tsconfig.json
@@ -2,7 +2,7 @@
     "extends": "expo/tsconfig.base",
     "compilerOptions": {
         "strict": true,
-        "baseUrl": "./",
+        "module": "esnext",
         "paths": {
             "@src/*": ["./src/*"],
             "@assets/*": ["./assets/*"]
